feat: fixed types

+++ Bank_System/Users/BankUser.cs
@@ -273,7 +273,7 @@ namespace Bank_System
             else
             {
                 decimal resultSum = Int32.Parse(sumForSend);
-                card.Transfer(cardForTranf, resultSum,Name,Common.Bank.FeeSending,Common.Bank.FeeReceipt);
+                card.Transfer(cardForTranf, resultSum,Name,(decimal)Common.Bank.FeeSending, (decimal)Common.Bank.FeeReceipt);
                 Message.SuccessMessage("Перевод успешно выполнен");
             }
         }
@@ -530,7 +530,7 @@ namespace Bank_System
                 {
                     if(transaction.SenderInitials == Name)
                     {
-                        resultSum += Common.Bank.FeeSending;
+                        resultSum += (double)Common.Bank.FeeSending;
 
                     }
                 }
@@ -550,7 +550,7 @@ namespace Bank_System
                 {
                     if(transaction.RecipientName == Name)
                     {
-                        resultSum += Common.Bank.FeeReceipt;
+                        resultSum += (double)Common.Bank.FeeReceipt;
                     }
                 }
             }
